Allow specifying of DNS resolvers, endpoint IP addresses (#6)
* progress

* remove test SOA work

* prep work for new requester

* all lookup options working

* some cmd util tests

* fix tests, add log lines, test cmd on CI

* CSVtoIPs test cases

* README updates

* change pad flag to `no-pad`, since we default to padding

* threadsafe dnscache

* haha

* endpoint parser tests

* dns cache tests

* LookupIP tests

* improve lookup tests

+++ provider.go
@@ -45,13 +45,13 @@ func (r DNSRR) RR() dns.RR {
 	// Build an RR header
 	rrhdr := dns.RR_Header{
 		Name:     r.Name,
-		Rrtype:   uint16(r.Type),
+		Rrtype:   r.Type,
 		Class:    dns.ClassINET,
-		Ttl:      uint32(r.TTL),
+		Ttl:      r.TTL,
 		Rdlength: uint16(len(r.Data)),
 	}
 
-	constructor, ok := dns.TypeToRR[uint16(r.Type)]
+	constructor, ok := dns.TypeToRR[r.Type]
 	if ok {
 		// Construct a new RR
 		rr = constructor()
